Create My_functions.cpp
I am getting rid of the folder, as I'm not really sure how to submit this assignment. I think, you're asking us to make another repository with similar files for gradescope or the repository from what you sent. I have to work in a little bit so If i can't finish I'll just have to come in